Create a box dropdown item list

Can be used to add dropdown items to a cardtool.

To insert in boxDropdown.

Usage

cardDropdown(..., icon = shiny::icon("wrench"))

cardDropdownItem(..., id = NULL, href = NULL, icon = NULL)

boxDropdown(..., icon = shiny::icon("wrench"))

boxDropdownItem(..., id = NULL, href = NULL, icon = NULL)

Arguments

...

Item content.

icon

Optional icon. Expect icon.

id

If passed, the item will behave like an action button.

href

Target url or page.

See also

Author

David Granjon, dgranjon@ymail.com