BS4 jumbotron for AdminLTE3

Create a jumbotron

bs4Jumbotron(
  ...,
  title = NULL,
  lead = NULL,
  href = NULL,
  btn_name = "More",
  status = c("primary", "warning", "danger", "info", "success")
)

Arguments

...
Any content.

title
Jumbotron title.

lead
Jumbotron lead.

href
Jumbrotron external link.

btn_name
Jumbotron button name.

status
Jumbotron background color. "primary", "success", "warning", "danger" or "info".

Examples

if(interactive()){
 library(shiny)
 library(bs4Dash)

 shiny::shinyApp(
   ui = bs4DashPage(
     navbar = bs4DashNavbar(),
     sidebar = bs4DashSidebar(),
     controlbar = bs4DashControlbar(),
     footer = bs4DashFooter(),
     title = "test",
     body = bs4DashBody(
      bs4Jumbotron(
      title = "Hello, world!",
      lead = "This is a simple hero unit, a simple jumbotron-style 
      component for calling extra attention to featured 
      content or information.",
      "It uses utility classes for typography and spacing 
      to space content out within the larger container.",
      status = "primary",
      href = "http://www.google.fr"
      )
     )
   ),
   server = function(input, output) {}
 )
}