Create a Boostrap 4 dashboard main sidebar menu item

Build an adminLTE3 dashboard main sidebar menu item

bs4SidebarMenuItem(
  text,
  ...,
  tabName = NULL,
  icon = NULL,
  expandedName = as.character(gsub("[[:space:]]", "", text)),
  startExpanded = FALSE,
  condition = NULL,
  selected = NULL
)

menuItem(
  text,
  ...,
  tabName = NULL,
  icon = NULL,
  expandedName = as.character(gsub("[[:space:]]", "", text)),
  startExpanded = FALSE,
  condition = NULL,
  selected = NULL
)

Arguments

text
Item name.

...
bs4SidebarMenuSubItem.

tabName
Should correspond exactly to the tabName given in bs4TabItem.

icon
Item icon.

expandedName
A unique name given to each menuItem that serves to indicate which one (if any) is currently expanded. (This is only applicable to menuItems that have children and it is mostly only useful for bookmarking state.)

startExpanded
Whether to expand the bs4SidebarMenuItem at start.

condition
When using bs4SidebarMenuItem with conditionalPanel, write the condition here (see https://github.com/RinteRface/bs4Dash/issues/35).

selected
If TRUE, this bs4SidebarMenuItem will start selected. If no item have selected=TRUE, then the first bs4SidebarMenuItem will start selected.

Note

See examples for a use case of the condition parameter.

Examples

if (interactive()) {
 # sidebarItem with conditional value
 library(shiny)
 library(bs4Dash)

 ui <- bs4DashPage(
  bs4DashNavbar(),
  bs4DashSidebar(
    bs4SidebarMenu(
      id = "sidebarMenu",
      bs4SidebarMenuItem(
        text = "Tab 1",
        tabName = "tab1"
      ),
      bs4SidebarMenuItem(
        condition = "input.show == true",
        text = "Tab 2",
        tabName = "tab2"
      )
    )
  ),
  bs4DashBody(
    bs4TabItems(
      bs4TabItem(
        tabName = "tab1",
        h1("Welcome!"),
        checkboxInput("show", "Show Tab 2", FALSE)
      ),
      bs4TabItem(
        tabName = "tab2",
        h1("Hey! You found me!")
      )
    )
   )
  )
  server <- function(input, output){}
  shinyApp(ui = ui, server = server)
}