Boostrap 4 column system

This function overwrites that of Shiny since there are differences between the Bootstrap 3 and Bootstrap 4 grid systems

column(width, ..., offset = 0)

Arguments

width

The grid width of the column (must be between 1 and 12.

...

Elements to include within the column.

offset

The number of columns to offset this column from the end of the previous column.

Examples

if (interactive()) {
 library(shiny)
 library(bs4Dash)
 
 ui <- bs4DashPage(
   navbar = bs4DashNavbar(),
   sidebar = dashboardSidebar(
     bs4SidebarMenu(
       bs4SidebarMenuItem(
         "Welcome!",
         tabName = "tab_welcome",
         icon = "home"
       )
     )
   ),
   
   body = bs4DashBody(
     bs4TabItems(
       bs4TabItem(
         tabName = "tab_welcome",
         fluidRow(
           column(
             width = 1,
             offset = 11,
             actionButton(
               "mybutton",label = "",icon = icon("circle-question")
             )
           )
           
         ),
         fluidRow(
           h2("Placeholder")
         )
       )
     )
   )
 )
 
 server <- function(input, output, session) {}
 shinyApp(ui = ui, server = server)
}