Boostrap 4 dashboard navbar

dashboardHeader creates an adminLTE3 dashboard navbar to be included in dashboardPage.

Usage

bs4DashNavbar(
  ...,
  title = NULL,
  titleWidth = NULL,
  disable = FALSE,
  .list = NULL,
  leftUi = NULL,
  rightUi = NULL,
  skin = "light",
  status = "white",
  border = TRUE,
  compact = FALSE,
  sidebarIcon = shiny::icon("bars"),
  controlbarIcon = shiny::icon("table-cells"),
  fixed = FALSE
)

dashboardHeader(
  ...,
  title = NULL,
  titleWidth = NULL,
  disable = FALSE,
  .list = NULL,
  leftUi = NULL,
  rightUi = NULL,
  skin = "light",
  status = "white",
  border = TRUE,
  compact = FALSE,
  sidebarIcon = shiny::icon("bars"),
  controlbarIcon = shiny::icon("table-cells"),
  fixed = FALSE
)

Arguments

...

Any UI element between left and right Ui. Can include navbarMenu to host the navigation in the navbar.

title

Dashboard title (displayed top-left side). Alternatively, use dashboardBrand for more evolved title.

titleWidth

This argument is deprecated; bs4Dash (AdminLTE3) title width is tightly related to the sidebar width, contrary to shinydashboard (AdminLTE2).

disable

If TRUE, don't display the header bar.

.list

An optional list containing items to put in the header. Same as the ... arguments, but in list format. This can be useful when working with programmatically generated items.

leftUi

Custom left Ui content. Any element like dropdownMenu.

rightUi

Custom right Ui content. Any element like dropdownMenu.

skin

Navbar skin. "dark" or "light".

status

Navbar status. Valid statuses are defined as follows:

primary: #007bff .

secondary: #6c757d .

info: #17a2b8 .

success: #28a745 .

warning: #ffc107 .

danger: #dc3545 .

gray-dark: #343a40 .

gray: #adb5bd .

white: #fff .

indigo: #6610f2 .

lightblue: #3c8dbc .

navy: #001f3f .

purple: #605ca8 .

fuchsia: #f012be .

pink: #e83e8c .

maroon: #d81b60 .

orange: #ff851b .

lime: #01ff70 .

teal: #39cccc .

olive: #3d9970 .

border

Whether to separate the navbar and body by a border. TRUE by default.

compact

Whether items should be compacted. FALSE by default.

sidebarIcon

Icon of the main sidebar toggle.

controlbarIcon

Icon to toggle the controlbar (left).

fixed

Whether to fix the navbar to the top. FALSE by default.

Author

David Granjon, dgranjon@ymail.com